+++ crates/forge-runner/src/running/target.rs
@@ -15,6 +15,7 @@ use cairo_lang_sierra::{
     ids::{ConcreteTypeId, FunctionId},
     program::{GenFunction, StatementIdx, TypeDeclaration},
 };
+use camino::Utf8Path;
 use rayon::iter::IntoParallelIterator;
 use rayon::iter::ParallelIterator;
 use std::{collections::HashMap, sync::Arc};
@@ -40,6 +41,7 @@ pub fn prepare_test_target(
     tracked_resource: &ForgeTrackedResource,
     name_filter: &NameFilter,
     partition_config: &PartitionConfig,
+    usc_cache_dir: &Utf8Path,
 ) -> Result<PrepareTestTargetResult> {
     let tests_location = test_target_raw.tests_location;
     let default_executables = vec![];
@@ -79,6 +81,7 @@ pub fn prepare_test_target(
 
     let casm_program = Arc::new(compile_raw_sierra_at_path(
         test_target_raw.sierra_program_path.as_std_path(),
+        Some(usc_cache_dir.as_std_path()),
     )?);
 
     let test_cases = matched_cases