+++ crates/forge/tests/integration/gas.rs
@@ -1174,7 +1174,7 @@ fn empty_test_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(0),
             l1_data_gas: GasAmount(0),
-            l2_gas: GasAmount(13_620),
+            l2_gas: GasAmount(5_910),
         },
     );
 }
@@ -1208,7 +1208,7 @@ fn declare_cost_is_omitted_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(0),
             l1_data_gas: GasAmount(0),
-            l2_gas: GasAmount(25_130),
+            l2_gas: GasAmount(17_420),
         },
     );
 }
@@ -1257,7 +1257,7 @@ fn deploy_syscall_cost_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(0),
             l1_data_gas: GasAmount(96),
-            l2_gas: GasAmount(183_710),
+            l2_gas: GasAmount(175_880),
         },
     );
 }
@@ -1297,7 +1297,7 @@ fn snforge_std_deploy_cost_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(0),
             l1_data_gas: GasAmount(96),
-            l2_gas: GasAmount(190_350),
+            l2_gas: GasAmount(182_520),
         },
     );
 }
@@ -1335,7 +1335,7 @@ fn keccak_cost_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(0),
             l1_data_gas: GasAmount(0),
-            l2_gas: GasAmount(226_407),
+            l2_gas: GasAmount(218_697),
         },
     );
 }
@@ -1370,7 +1370,7 @@ fn sha512_cost_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(0),
             l1_data_gas: GasAmount(0),
-            l2_gas: GasAmount(2_436_110),
+            l2_gas: GasAmount(2_428_400),
         },
     );
 }
@@ -1421,7 +1421,7 @@ fn contract_keccak_cost_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(0),
             l1_data_gas: GasAmount(96),
-            l2_gas: GasAmount(1_352_655),
+            l2_gas: GasAmount(1_344_825),
         },
     );
 }
@@ -1475,7 +1475,7 @@ fn storage_write_cost_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(0),
             l1_data_gas: GasAmount(192),
-            l2_gas: GasAmount(740_950),
+            l2_gas: GasAmount(733_120),
         },
     );
 }
@@ -1529,7 +1529,7 @@ fn multiple_storage_writes_cost_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(0),
             l1_data_gas: GasAmount(192),
-            l2_gas: GasAmount(897_590),
+            l2_gas: GasAmount(889_760),
         },
     );
 }
@@ -1591,7 +1591,7 @@ fn l1_message_cost_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(29_524),
             l1_data_gas: GasAmount(96),
-            l2_gas: GasAmount(292_710),
+            l2_gas: GasAmount(284_880),
         },
     );
 }
@@ -1655,7 +1655,7 @@ fn l1_message_cost_for_proxy_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(29_524),
             l1_data_gas: GasAmount(192),
-            l2_gas: GasAmount(556_020),
+            l2_gas: GasAmount(548_190),
         },
     );
 }
@@ -1703,7 +1703,7 @@ fn events_cost_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(0),
             l1_data_gas: GasAmount(0),
-            l2_gas: GasAmount(205_190),
+            l2_gas: GasAmount(197_480),
         },
     );
 }
@@ -1750,7 +1750,7 @@ fn events_contract_cost_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(0),
             l1_data_gas: GasAmount(96),
-            l2_gas: GasAmount(470_720),
+            l2_gas: GasAmount(462_890),
         },
     );
 }
@@ -1823,7 +1823,7 @@ fn nested_call_cost_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(0),
             l1_data_gas: GasAmount(288),
-            l2_gas: GasAmount(1_938_482),
+            l2_gas: GasAmount(1_930_652),
         },
     );
 }
@@ -1896,7 +1896,7 @@ fn nested_call_cost_in_forked_contract_sierra_gas() {
         GasVector {
             l1_gas: GasAmount(0),
             l1_data_gas: GasAmount(192),
-            l2_gas: GasAmount(1_810_912),
+            l2_gas: GasAmount(1_803_082),
         },
     );
 }