+++ crates/sncast/src/starknet_commands/multicall/mod.rs
@@ -60,7 +60,7 @@ pub async fn multicall(
         }
     }
 
-    match &multicall.command {
+    match multicall.command {
         starknet_commands::multicall::Commands::New(new) => {
             if let Some(output_path) = &new.output_path {
                 let result = starknet_commands::multicall::new::write_empty_template(
@@ -84,7 +84,7 @@ pub async fn multicall(
             let account = get_account(&config, &provider, &run.rpc, ui).await?;
             let result = with_account!(&account, |account| {
                 starknet_commands::multicall::run::run(
-                    run.clone(),
+                    run,
                     account,
                     &provider,
                     &config,
@@ -105,11 +105,12 @@ pub async fn multicall(
         }
         starknet_commands::multicall::Commands::Execute(execute) => {
             let provider = execute.rpc.get_provider(&config, ui).await?;
+
             let account = get_account(&config, &provider, &execute.rpc, ui).await?;
 
             let result = with_account!(&account, |account| {
                 starknet_commands::multicall::execute::execute(
-                    *execute.clone(),
+                    *execute,
                     account,
                     &provider,
                     &config,