+++ crates/sncast/tests/e2e/show_config.rs
@@ -37,7 +37,7 @@ async fn test_show_config_json() {
     let snapbox = runner(&args).current_dir(tempdir.path());
 
     snapbox.assert().success().stdout_eq(indoc! {r#"
-        {"account":"user1","accounts_file_path":"../account-file","alias_count":0,"block_explorer":"Voyager","chain_id":"alpha-sepolia","command":"show-config","keystore":null,"network":null,"networks":{"devnet":null,"mainnet":null,"sepolia":null},"profile":null,"rpc_url":"http://127.0.0.1:5055/rpc","scarb_profile":"release","show_explorer_links":true,"type":"response","wait_retry_interval":5,"wait_timeout":300}
+        {"account":"user1","accounts_file_path":"../account-file","alias_count":0,"block_explorer":"Voyager","chain_id":"alpha-sepolia","command":"show-config","fee_params":null,"keystore":null,"network":null,"networks":{"devnet":null,"mainnet":null,"sepolia":null},"profile":null,"rpc_url":"http://127.0.0.1:5055/rpc","scarb_profile":"release","show_explorer_links":true,"type":"response","wait_retry_interval":5,"wait_timeout":300}
     "#});
 }
 
@@ -194,6 +194,30 @@ async fn test_show_config_with_network() {
         Scarb Profile:       release
         Alias Count:         0
         Sepolia URL:         http://127.0.0.1:5055/rpc
+    "});
+}
+
+#[tokio::test]
+async fn test_show_config_with_fee() {
+    let tempdir = copy_config_to_tempdir("tests/data/files/snfoundry_correct.toml", None);
+    let args = vec!["--profile", "with_fee", "show-config"];
+
+    let snapbox = runner(&args).current_dir(tempdir.path());
+
+    snapbox.assert().success().stdout_eq(formatdoc! {r"
+        Profile:             with_fee
+        Chain ID:            alpha-sepolia
+        RPC URL:             http://127.0.0.1:5055/rpc
+        Account:             user1
+        Accounts File Path:  ../account-file
+        Wait Timeout:        300s
+        Wait Retry Interval: 5s
+        Show Explorer Links: true
+        Block Explorer:      Voyager
+        Scarb Profile:       release
+        Alias Count:         0
+        L2 Gas:              1000000
+        Estimate Tip:        true
     "});
 }